Renamed start task and backend/frontend entry points
